move actors to it's own package
